Feature/giyoung

+++ src/main/java/com/software/graduable/grade/dto/GradeOutputDTO.java
@@ -23,7 +23,7 @@ public class GradeOutputDTO {
     // 과목명
     private String courseName;
     // 학점(설계)
-    private int credit;
+    private Double credit;
     // 성적
     private String grade;
     // 비고
